DELETE /api/analyses/:id #138
