+++ flask/app/__init__.py
@@ -36,5 +36,6 @@ def create_app(config):
         from . import manage
         from . import routes
         from . import buckets
+        from . import analyses
 
         return app
